+++ ui/src/app/url-list/url-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ApiService, URLItem } from './api.service';
+import { ApiService, URLItem } from '../api.service';
 
 @Component({
   selector: 'app-url-list',
